Some fairly serious changes to folder builder, so this bumps the version to 0.8.0.  Still working on fixing unit tests on the server.
This change the concept of shared folders which we may or may not want to undo.  Any user
can share a folder, and then unshare it.  So there isn't a separate heirarchy of public
folders.  A shared folder, should kinda become a child of the root for everyone, except
for the person sharing it, in which case it exists somewhere in their hierarchy.

So, showing the first level of folders is basically pulling all children of the root
with your user ID, and all folders not with your ID that are shared.

All subfolders of a shared folder are implicitly shared.

Then we can also extend this to allow people to share folders only with certain other people.