Udpated to SDK 0.11.0 which changes keyword confidence to a value between 0 and 1.  However, since the KeywordSchema defines how keywords are structured in elastic, that information has be to be exposed for SearchService to do its work, so implementation is not hidden.  How the data is structured on on asset is currently public.
Future course of action could be an elastic plugin or moving moving all the knowledge of the final structure of
keywords ito the Archivist.  So the KeywordSchema would store the keyword+confidence in something like
a Map<String, Double>, and then we would use our custom JSON serializer to bucket the words at json serialization time.
This would be "just in time" bucketing vs pre-bucketing like we do now.