This patch does a few intertwined things which is part of a slow evolution of handling other types of files besides images, and what to do when they fail.
With the ingests, want to ensure that we don't pile everything into (Schema)AssetMetadataProcessor for every type, so
I've moved all its logic into ImageProcessor and SchemaAssetMetadataProcessor determines the data is a image
and sends it off to ImageProcessor.  I think going forward it might be useful to say AssetMetadataProcessor always runs and
it can't be taken out of a pipeline.

2. For source data thats broken in some way, I'd like to handle them in a specific area of the code, maybe once
the ingest is done.  For now I throw  an IngestProcessorException of the image is corrupted or can't be opened
and log that.  In other cases where the metadata is fine but subsequent processinf fails, it still uses the same
errors field.