Room Selections
Changed to support atomic propagation of shared room state.

- AssetFilter: no longer has a selected property.  The selected of a room
are a property of a room.  To get just the IDs of the selected
assets use the /api/v1/rooms/current/selection endpoint.

- AssetSearchBuilder: roomId was removed, added useAsRoomSearch propery
which when set to true (defaults to true) will set the search as the
current roo search and emit the search.

-MessageType: - ASSET_SEARCH, ASSET_SELECT, ASSET_DESELECT are deprecated
in favour of ROOM_SELECTION_UPDATE and ROOM_SEARCH_UPDATE.

- RoomBuilder: its now possible to specify a selection and search when a
room is build, so if a person creates a new room their current selection
and search will be what people get when the join.

- SharedRoomState: A class that holds the current shared state of a room.
Currently this includes a search and a selection.

- RoomService: updated with methods to support the new room concepts.