Room selection refactor.
This changes the room selection to be a property of the room and not of the
asset itself.

Key changes:

- storing selections in elastic generates many dirty lucene segments which have to be
cleaned up and compacted.  So, selections are stored in SQL.

- Selections are atomic.  If N selections happen simultanenously, the locking around
the SQL row ensures that they happen serially.

- The REST endpoints for the room concerning setting/getting selections don't require
the client to know the roomId, to avoid having the client libs have to supply
something the sever already knows.

- The message types for selection/search have changed, see SDK for notes.

- The endpoints for setting and getting the selections are in RoomController.  There
is a convinience method which returns a SharedRoomState which lets you obtain all
the necessary data upon joinin a room with a single API call.

Future improvements:

- Currently the API exects the whole selection to be passed around, rather than changes
to the selection.  This is slightly more complex to implement but can be done if necessary.