Ingestors: AssetMetadataProcessor is now default.
AssetMetadataProcessor basically serves as facade processor which executes processors
that handle specific asset types like, ImageProcessor.  Eventually we'll just add an
ingest pipeline flag that determines which Asset types to target.  The types are defined
by the AssetType enum in the SDK, which is currently:

Image, Video, Document, Model and Unknown.

The old processors were removed and the Schema based processed renamed.  This patch also
stores the processor data as Json rather than java serialization which is not
human readable.