AssetBuilder gets an InputStream and a BufferedImage.
This solves two issues.

a. pretty much all java APIs that read something off disk take an InputStream object.  However,
most input streams are not resetable, meaning you can't reuse them.  AssetBuilder now has a special
type of InputStream which is reusable, which cuts back the number of times the file has to be
read from disk to 1.

b.  The buffered image.  Not all assets are images, and so the job of some processors is to extract
some kind of image from the data and set this value.  For ImageIngestor its easy, for some
other types of data its slightly more invovled.  This allows us to reuse proxy processor for all
types of data.