Ingest: the next evoltuion.
This patch to ingests opens the door to handling other formats besides just images.  Each ingestor
sets the file formats it supports, which in turn are used to filter both the files that get
sent to the asset threads, and which processors get run.

This patch also utilizes new properties on assetbuilder which bring some performance enhancements