@103691874 Local cloud caches
The ObjectFileSystem class is the interface that the Archivist use to store its data,
be it proxies or data obtained from any external source.  The UUIDFileSystemImpl is
an implementation of this interface which uses UUIDs to avoid naming collions and provides
and easy way to shard files accross servers if we do a distributed version.

This system replaces ImageService, since all it was doing was providing directories to store image.