Giant reorg which aims to move the whole processing pipeline into a separate process.
The main side effect to this patch is that its not really possible to ingest assets by the traditional means.
So pretty much all the search tests fail.

Anthoer side effect effect here is that some classes were pulling in transative dependencies which no longer exist.  There was almost
always a better alternative which I've used.  That is why so many files area being touched.