Notes: simple note feature
This is a simple note feature, a lot more left to do.  You can attach a single note to multiple assets,
I'm not even sure if that is useful, and it kinda adds more complexity to it.  It would be great to
display somehow when a note references multiple things.  We'll also need a way to basically split
the note in case you add a note to 3 assets then decide you want to edit just the note for one and not all.

This gets me enough so I can replicate notes between archivists.