The great Document/Attribute refactor.
This patch allows us to set attributes more than 2 levels deep.  The biggest change here is
the attribute delimited, which was a ".", and is now a ":".  This is because the "." is
a special elasticsearch character.

Under the hood, there are 2 ways an attribute can be set.  The attribute can be a direct
property of the class, it can be an arbitrary attribute in a Map<String,Object>