Keywords: simplified keywords system.
- removed confidence for now. The point of the float level confidence was to hide the implemnation,
but since the client is directly exposed to the JSON document and the API they are exposed to the implemenation
of confidence.  A better route is probably to let people manage it on their own by allowing them
to query only certain fields.  For example they could but low confidence words in a separate field.

- added methods called addSuggestKeywords, the boolean in addKeywords was annoying and in many
places people commneted on what it was actually used for.  So, this is more clear.