Archvist: chanages to folder REST endpoints.
Getting a folder by path, rather than _, using _path to differentiate, somewhat more clear.

-    @RequestMapping(value="/api/v1/folders/_/**", method=RequestMethod.GET)
+    @RequestMapping(value="/api/v1/folders/_path/**", method=RequestMethod.GET)

Is now the proper way of getting child folders
@RequestMapping(value="/api/v1/folders/{id}/folders", method=RequestMethod.GET)

Endpoint is removed since it really has to be implemented on the client side to allow for paging.
@RequestMapping(value="/api/v1/folders/{id}/assets", method=RequestMethod.GET)

Added a new endpoint for getting a child folder.
@RequestMapping(value="/api/v1/folders/{id}/folders/{name})