Add filebeat shipper for forwarding docker logs. (#229)
EVI-1410
This works well on linux and from the research I've done should also
work on mac osx, but I've not tested it.  For non-production K8s will
need a slightly different config.  For production will need a more
robust set of services or maybe better to use a sass provider like
datadog.