Serve proxies from _stream if specific types are requestd.
If nothing special is requested through an Accept header or an `type`
query parameter, behavior is unchanged (original source will be streamed
if the user can export the asset and does not request a specific proxy type of `image` or `video`).

This should fix video playback without breaking anything else.
Transcoded video files should be played in curator instead of the source
asset.