Phase 1 of refactoring content negotiation on _stream endpoint.
Everything is more or less based on accept headers now, except for the case where the 'ext' query
string parameter is used. If no accept headers are sent, then */* is assumed.  In this
case a proxy of the same overall type (video, etc) will be selected first, then otherwise
fall back to an image.

- Removed unused query parameter 'type', which was not used by Curator.
- Added 'ext' query parameter back in which is used by Curator for vide playback.
- Normalized the accept header and the 'ext' param into a list of acceptable media types
  so less HTTP details leaking into service layer.
- Replaced manual Accept header parsig with spring HttpHeaders